[MPI] GetOrderArray automatically calls the MPI based order computation when needed, ArrayPrecondition not manually needed
